Update US07_USA_Pluto.m3u
I finally cracked the code, for real this time. 😀
So the thing is that Pluto TV recently made some changes to their API and their ads are now embedded into the stream itself.
This means that when adding the "advertisementId" and "serverSideAds" flags to the url, the ads would overlap and therefore crash the streams.
By reading through the API, I also learned that you don't have to manually set the "sid" value. Set it to "channel" and it'll pick up the correct value automatically.